Fuzzer Icon Update
- updates sidedrawer warning icon
- some code reorganization
