+++ ui/src/components/Icon/IconTemplates.js
@@ -129,6 +129,19 @@ const IconTemplates = () => (
                     1 1-1h1a1 1 0 0 1 1 1v2zM5 7a1 1 0 0 0-1 1v9a1 1 0 0 0 1 1h13a1 1 0 0 0 1-1V8a1 1 0 0 0-1-1H5zm9 5h2a1 1 0 0 1 1 1v2a1 1 0 0
                     1-1 1h-2a1 1 0 0 1-1-1v-2a1 1 0 0 1 1-1z"/>
             </symbol>
+            <symbol id={ICON_NAMES.EXCLAMATION_MARK_ROUND} viewBox="0 0 22 22">
+                <path fillRule="evenodd" clipRule="evenodd" d="M2 11C2 6.029 6.0294 2 11 2C15.9706 2 20 6.029 20 11C20 15.971 15.9706 20 11 20C6.0294 20 2
+                    15.971 2 11ZM4 11C4 7.134 7.134 4 11 4C14.866 4 18 7.134 18 11C18 14.866 14.866 18 11 18C7.134 18 4 14.866 4 11ZM10 7C10 6.448 10.4477 6
+                    11 6C11.5523 6 12 6.448 12 7L12 12C12 12.552 11.5523 13 11 13C10.4477 13 10 12.552 10 12V7ZM11 16C11.5523 16 12 15.552 12 15C12 14.448
+                    11.5523 14 11 14C10.4477 14 10 14.448 10 15C10 15.552 10.4477 16 11 16Z" />
+                <mask id="warning-round-mask" maskUnits="userSpaceOnUse" x="2" y="2" width="18" height="18">
+                    <path fillRule="evenodd" clipRule="evenodd" d="M2 11C2 6.029 6.0294 2 11 2C15.9706 2 20 6.029 20 11C20 15.971 15.9706 20 11 20C6.0294
+                        20 2 15.971 2 11ZM4 11C4 7.134 7.134 4 11 4C14.866 4 18 7.134 18 11C18 14.866 14.866 18 11 18C7.134 18 4 14.866 4 11ZM10 7C10 6.448
+                        10.4477 6 11 6C11.5523 6 12 6.448 12 7L12 12C12 12.552 11.5523 13 11 13C10.4477 13 10 12.552 10 12V7ZM11 16C11.5523 16 12 15.552 12
+                        15C12 14.448 11.5523 14 11 14C10.4477 14 10 14.448 10 15C10 15.552 10.4477 16 11 16Z" fill="white"/>
+                </mask>
+                <g mask="url(#warning-round-mask)"></g>
+            </symbol>
             <symbol id={ICON_NAMES.FILTER} viewBox="0 0 18 14">
                 <path fillRule="evenodd" clipRule="evenodd" d="M0.816406 1.54552C0.816406 0.792794 1.42732 0.181885 2.18004 0.181885H15.8164C16.5691
                     0.181885 17.18 0.792794 17.18 1.54552C17.18 2.29825 16.5691 2.90916 15.8164 2.90916H2.18004C1.42732 2.90916 0.816406 2.29825
